Added a new Command(s) and Handler(s)
The bot is now able to response to messages which contains its name. He
uses a variety of answers randomly. Also does it response with a set of
answers to bodo. Further the bot can give a user a role to post links
and pics within a minute, but this command only works for mods.