Big update
The bot is now able to:
*) react to some welcome messages
*) congratulate to a role upgrade

Also added more tests and changed the user got mention class. The log4j
property file is now outside of the .jar so I can change the log level
during runtime. General changes in logging, so that the log file gets
filled with the appropriate information.